+++ apstools/utils.py
@@ -3,6 +3,7 @@ Various utilities
 
 .. autosummary::
    
+   ~becplot_prune_fifo
    ~cleanupText
    ~command_list_as_table
    ~connect_pvlist
@@ -74,6 +75,45 @@ MAX_EPICS_STRINGOUT_LENGTH = 40
 
 
 class ExcelReadError(xlrd.XLRDError): ...
+
+
+def becplot_prune_fifo(n, y, x):
+    """
+    find the plot with axes x and y and replot with only the last *n* lines
+
+    Note: this is not a bluesky plan.  Call it as normal Python function.
+
+    EXAMPLE::
+
+        becplot_prune_fifo(1, noisy, m1)
+
+    PARAMETERS
+    
+    n : int
+        number of plots to keep
+    
+    y : object
+        ophyd Signal object on dependent (y) axis
+    
+    x : object
+        ophyd Signal object on independent (x) axis
+    
+    Assumes object `bec` (instance of BestEffortCallback) is defined.
+    """
+    global bec
+    for liveplot in bec._live_plots.values():
+        lp = liveplot.get(y.name)
+        if lp is None:
+            logging.debug(f"no LivePlot with name {y.name}")
+            continue
+        if lp.x != x.name or lp.y != y.name:
+            logging.debug(f"no LivePlot with axes ('{x.name}', '{y.name}')")
+            continue
+        if len(lp.ax.lines) > n:
+            logging.debug(f"limiting LivePlot({y.name}) to {n} traces")
+            lp.ax.lines = lp.ax.lines[-n:]
+            lp.update_plot()
+
 
 
 def cleanupText(text):